Responsive design for 3(4) viewing sizes
More less I configured responsive viewing sizes with @media screen
queries for all pages and for 3 sizes. I don't have idea how to do it
for 767-480, so I did it only for index - and it looks terrible :)
Please take a look at it and feel free to criticise my ideas, and change
whatever you think that should be changed. I think this is only a raw
version, and we have to refine it.