Added swiping navigation
Pages are looped in following order index->weekly->reviewers->index. To
nawigate we are swiping left or right. I wanted to make two fingers
swype instead of one finger swype, however... I don't if it is only iPad
problem but two finger swyping works up/down/left but not right!! We
cannot fully exclude triggering accidental page navigation action while
moving panorama, but the times are configured in a way that it is not
quite likely probable.